chore: upgrade dependencies (#107)
* chore: upgrade dependencies

* chore: fix types

* chore: upgrade zod

* fix: types

* fix: build

* chore: remove resolutions

* chore: add logging

* fix: add sailing to properly validate max total level

* fix: types

* chore: fix dirty field preventing submission

+++ apps/web/config/constants.client.ts
@@ -13,7 +13,7 @@ const ClientConfigSchema = z.object({
   wiki: z.object({
     baseUrl: z.literal('https://oldschool.runescape.wiki'),
     userAgent: z.literal('Irons-Tavern-Rank-Calculator (Discord @joemckie)'),
-    queryLimit: z.number().int().min(1).max(5000),
+    queryLimit: z.number().min(1).max(5000),
   }),
   wikiSync: z.object({
     baseUrl: z.literal('https://sync.runescape.wiki'),
